代理模式的示例
