+++ src/com/hnxyhcz/designPatterns/ProxyPatterns/ClientMain.java
@@ -1,5 +1,12 @@
 package com.hnxyhcz.designPatterns.ProxyPatterns;
 
+import com.hnxyhcz.designPatterns.ProxyPatterns.proxy.Proxy;
+
+/**
+ * 代理模式：张三让戴利替他向李娇娇送礼物
+ * @author hnxyhcz
+ *
+ */
 public class ClientMain {
 
 	public static void main(String[] args) {
