+++ src/com/hnxyhcz/designPatterns/ProxyPatterns/SchoolGirl.java
@@ -1,5 +1,10 @@
 package com.hnxyhcz.designPatterns.ProxyPatterns;
 
+/**
+ * 姑凉
+ * @author hnxyhcz
+ *
+ */
 public class SchoolGirl {
 
 	private String Name;
