+++ src/com/hnxyhcz/designPatterns/ProxyPatterns/subject/GiveGift.java
@@ -1,5 +1,10 @@
-package com.hnxyhcz.designPatterns.ProxyPatterns;
+package com.hnxyhcz.designPatterns.ProxyPatterns.subject;
 
+/**
+ * 抽象对象角色
+ * @author hnxyhcz
+ *
+ */
 public interface GiveGift {
 	
 	public void GiveDolls();
